Refactor useConfig imports: update to use new hook from hooks directory
