+++ apps/web/src/pages/DashboardPage.tsx
@@ -3,7 +3,7 @@ import { useAuth } from '../store/authUtils';
 import { useProfile } from '../hooks/useProfile';
 import { useUserRegistrations } from '../hooks/useUserRegistrations';
 import { useCampRegistration } from '../hooks/useCampRegistration';
-import { useConfig } from '../store/ConfigContextDefinition';
+import { useConfig } from '../hooks/useConfig';
 import { getFriendlyDayName, formatTime } from '../utils/shiftUtils';
 import { isRegistrationAccessible, getRegistrationStatusMessage, getActiveRegistrations, getCancelledRegistrations } from '../utils/registrationUtils';
 import { PATHS } from '../routes';
