+++ apps/web/src/pages/registration/RegistrationPage.test.tsx
@@ -7,7 +7,7 @@ import * as useRegistrationModule from '../../hooks/useRegistration';
 import * as useCampingOptionsModule from '../../hooks/useCampingOptions';
 import * as useProfileModule from '../../hooks/useProfile';
 import * as useCampRegistrationModule from '../../hooks/useCampRegistration';
-import * as useConfigModule from '../../store/ConfigContext';
+import * as useConfigModule from '../../hooks/useConfig';
 
 // Mock modules
 vi.mock('../../hooks/useRegistration');
@@ -261,6 +261,9 @@ describe('RegistrationPage', () => {
       isLoading: false,
       error: null,
       refreshConfig: vi.fn(),
+      isConnecting: false,
+      isConnected: true,
+      connectionError: null,
     });
   });
 
@@ -282,7 +285,10 @@ describe('RegistrationPage', () => {
             logout: vi.fn(), 
             isLoading: false, 
             error: null, 
-            isAuthenticated 
+            isAuthenticated,
+            isConnecting: false,
+            isConnected: true,
+            connectionError: null,
           }}
         >
           <RegistrationPage />
